requirements.txt

langchain==0.1.16
langgraph==0.0.32
langchain-openai==0.0.8
pinecone-client==3.0.0
python-dotenv==1.0.0
redis>=4.5.5
jupyterlab
